--- a/qiskit/providers/models/pulsedefaults.py
+++ b/qiskit/providers/models/pulsedefaults.py
@@ -106,7 +106,7 @@ class Command:
         Args:
             name (str): The name of the command
             qubits: The qubits for the command
-            sequence (List[PulseQobjInstruction]): The sequence for the Command
+            sequence (PulseQobjInstruction): The sequence for the Command
             kwargs: Optional additional fields
         """
         self._data = {}