--- a/qiskit/qobj/pulse_qobj.py
+++ b/qiskit/qobj/pulse_qobj.py
@@ -226,20 +226,34 @@ class PulseQobjInstruction:
         Returns:
             PulseQobjInstruction: The object from the input dictionary.
         """
-        t0 = data.pop("t0")
-        name = data.pop("name")
-        if "kernels" in data:
-            kernels = data.pop("kernels")
-            kernel_obj = [QobjMeasurementOption.from_dict(x) for x in kernels]
-            data["kernels"] = kernel_obj
-        if "discriminators" in data:
-            discriminators = data.pop("discriminators")
-            discriminators_obj = [QobjMeasurementOption.from_dict(x) for x in discriminators]
-            data["discriminators"] = discriminators_obj
-        if "parameters" in data and "amp" in data["parameters"]:
-            data["parameters"]["amp"] = _to_complex(data["parameters"]["amp"])
+        schema = {
+            "discriminators": QobjMeasurementOption,
+            "kernels": QobjMeasurementOption,
+        }
+        skip = ["t0", "name"]
 
-        return cls(name, t0, **data)
+        # Pulse instruction data is nested dictionary.
+        # To avoid deepcopy and avoid mutating the source object, create new dict here.
+        in_data = {}
+        for key, value in data.items():
+            if key in skip:
+                continue
+            if key == "parameters":
+                # This is flat dictionary of parametric pulse parameters
+                formatted_value = value.copy()
+                if "amp" in formatted_value:
+                    formatted_value["amp"] = _to_complex(formatted_value["amp"])
+                in_data[key] = formatted_value
+                continue
+            if key in schema:
+                if isinstance(value, list):
+                    in_data[key] = list(map(schema[key].from_dict, value))
+                else:
+                    in_data[key] = schema[key].from_dict(value)
+            else:
+                in_data[key] = value
+
+        return cls(data["name"], data["t0"], **in_data)
 
     def __eq__(self, other):
         if isinstance(other, PulseQobjInstruction):