… test/python/compiler/test_transpiler.py
… test/python/pulse/test_calibration_entries.py