<!--
⚠️ If you do not respect this template, your pull request will be closed.
⚠️ Your pull request title should be short detailed and understandable for all.
⚠️ Also, please add a release note file using reno if the change needs to be
  documented in the release notes.
⚠️ If your pull request fixes an open issue, please link to the issue.

- [ ] I have added the tests to cover my changes.
- [ ] I have updated the documentation accordingly.
- [ ] I have read the CONTRIBUTING document.
-->

### Summary

Loading backend `PulseDefaults` with 100+ qubit is significantly slow because it immediately converts all pulse qobjs into `Schedule`s when it is called for the first time. This induces significant performance regression in transpile because all preset passes require the default object, i.e. `InstructionScheduleMap`, to invoke pulse gate pass for user calibration.

This PR allows us to convert it as-needed basis for speed up.

Fix https://github.com/Qiskit/qiskit-terra/issues/7914

Blocked by https://github.com/Qiskit/qiskit-terra/pull/8839

### Details and comments

This PR allows `InstructionScheduleMap` to take unconverted pulse qobj, and invoke the conversion logic when the `InstructionScheduleMap.get` method is called for a particular calibration. 

This is mainly done by the commit 64a37122e64d7cb98b75cc79a8574759460592d0

![image](https://user-images.githubusercontent.com/39517270/195410961-6e6f383a-74d0-4a07-9f3f-f241fbdca879.png)

According to the profile, I also removed the use of deep copy (this is added because Qobj data is a nested dict and to avoid mutating it) in the commit 0e1b5f2ef5756e5ba625e21e0c95d5372a86af02 (roughly 3x speedup)

![image](https://user-images.githubusercontent.com/39517270/195411235-05dfa39d-52bc-4820-890e-2b14679b006f.png)

This is output of airspeed velocity test:

```
Benchmarks that have improved:

       before           after         ratio
     [e9736b41]       [0e1b5f2e]
     <fix/pulse_qobj_converter>       <upgrade/lazy_instmap_qobj_conversion>
-        774±10ms       21.0±0.5ms     0.03  pulse.load_pulse_defaults.PulseDefaultsBench.time_building_defaults(1000)
-        88.1±2ms      2.09±0.08ms     0.02  pulse.load_pulse_defaults.PulseDefaultsBench.time_building_defaults(100)
-      14.8±0.6ms          292±6μs     0.02  pulse.load_pulse_defaults.PulseDefaultsBench.time_building_defaults(10)
-      6.66±0.3ms          112±3μs     0.02  pulse.load_pulse_defaults.PulseDefaultsBench.time_building_defaults(0)

Benchmarks that have stayed the same:

       before           after         ratio
     [e9736b41]       [0e1b5f2e]
     <fix/pulse_qobj_converter>       <upgrade/lazy_instmap_qobj_conversion>
         17.6±1ms       18.4±0.6ms     1.05  pulse.load_pulse_defaults.CircuitSchedulingBench.time_scheduling_circuits(3)
         74.8±2ms         73.0±3ms     0.98  pulse.load_pulse_defaults.CircuitSchedulingBench.time_scheduling_circuits(15)

Benchmarks that have got worse:

       before           after         ratio
     [e9736b41]       [0e1b5f2e]
     <fix/pulse_qobj_converter>       <upgrade/lazy_instmap_qobj_conversion>
+      6.74±0.3ms       9.98±0.3ms     1.48  pulse.load_pulse_defaults.CircuitSchedulingBench.time_scheduling_circuits(1)
+      12.6±0.7ms         16.0±2ms     1.27  pulse.load_pulse_defaults.CircuitSchedulingBench.time_scheduling_circuits(2)
```


As you can see we achieved roughly x50 speedup with these two commits. Note that scheduling test suffers the performance regression as a side effect of the lazy conversion (because it internally `.get` schedule). I guess the same problem will occur in the V2 converter https://github.com/Qiskit/qiskit-terra/pull/8759, but the converter should be able to avoid the performance regression by implementing the similar lazy conversion mechanism. Regarding circuit scheduler, in principle, we should improve the mechanism to build `Schedule`, as proposed in https://github.com/Qiskit/qiskit-terra/pull/8029.