@mtreinish 2022-10-12
> I guess the same problem will occur in the V2 converter https://github.com/Qiskit/qiskit-terra/pull/8759, but the converter should be able to avoid the performance regression by implementing the similar lazy conversion mechanism.

All the IBM provider (except for the one with the q) backends will have this same problem too since they all pull in the pulse schedules from defaults for `InstructionProperties` during the `Target` generation