@nkanazawa1989 2022-10-12
> All the IBM provider (except for the one with the q) backends will have this same problem too since they all pull in the pulse schedules from defaults for InstructionProperties during the Target generation

Do they have own mechanism for Target construction or just using some code in terra? If they rely on terra code I can update InstructionProperties in follow-up/in this PR.