@mtreinish 2022-10-12
It's a copy and paste of the same basic code that's in the `BackendV2Converter`, which was all based on a rough draft that I wrote about ~8 months ago for taking the IBM JSON response payloads and creating a `Target` from the dictionaries :

- https://github.com/Qiskit/qiskit-ibm-provider/blob/main/qiskit_ibm_provider/utils/backend_converter.py
- https://github.com/Qiskit/qiskit-ibm-runtime/blob/main/qiskit_ibm_runtime/utils/backend_converter.py