@nkanazawa1989 2022-10-13
Added new commit https://github.com/Qiskit/qiskit-terra/pull/8885/commits/26903089dcf10120e3f1c8b5e0accdcf9c8256d4 to support lazy conversion for V2. This commit allows `InstructionProperties` to take `PulseQobjDef` instance in addition to `Schedule` and `ScheduleBlock`. Then the conversion logic is invoked when `InstructionProperties.calibration` is called.