@nkanazawa1989 2023-01-12
Thanks for catching the missing release note. I had forgotten to write the note. Now it's added.